Fixes end-of-life inputs
The end-of-life was incorrectly modelled.
It relied on "treatment of used passenger car", from ecoinvent, which is in fact almost empty.
Hence, it needs to rely separately on "treatment of used glider", "treatment of used powertrain" and "treatment of used engine" separately. This adds about 5 to 6g CO2-eq./vkm.